+++ Specs/Core/Matrix3Spec.js
@@ -161,6 +161,51 @@ defineSuite([
         expect(returnedResult).toEqualEpsilon(expected, CesiumMath.EPSILON15);
     });
 
+    it('fromHeadingPitchRoll works without a result parameter', function() {
+        var sPiOver4 = Math.sin(CesiumMath.PI_OVER_FOUR);
+        var cPiOver4 = Math.cos(CesiumMath.PI_OVER_FOUR);
+        var sPiOver2 = Math.sin(CesiumMath.PI_OVER_TWO);
+        var cPiOver2 = Math.cos(CesiumMath.PI_OVER_TWO);
+
+        var tmp = Cartesian3.multiplyByScalar(new Cartesian3(0.0, 0.0, 1.0), sPiOver4, new Cartesian3());
+        var quaternion = new Quaternion(tmp.x, tmp.y, tmp.z, cPiOver4);
+        var headingPitchRoll = HeadingPitchRoll.fromQuaternion(quaternion, undefined, true);
+        var expected = new Matrix3(cPiOver2, -sPiOver2, 0.0, sPiOver2, cPiOver2, 0.0, 0.0, 0.0, 1.0);
+
+        var returnedResult = Matrix3.fromHeadingPitchRoll(headingPitchRoll, undefined, true);
+        expect(returnedResult).toEqualEpsilon(expected, CesiumMath.EPSILON15);
+    });
+
+    it('fromHeadingPitchRoll works with a result parameter', function() {
+        var sPiOver4 = Math.sin(CesiumMath.PI_OVER_FOUR);
+        var cPiOver4 = Math.cos(CesiumMath.PI_OVER_FOUR);
+        var sPiOver2 = Math.sin(CesiumMath.PI_OVER_TWO);
+        var cPiOver2 = Math.cos(CesiumMath.PI_OVER_TWO);
+
+        var tmp = Cartesian3.multiplyByScalar(new Cartesian3(0.0, 0.0, 1.0), sPiOver4, new Cartesian3());
+        var quaternion = new Quaternion(tmp.x, tmp.y, tmp.z, cPiOver4);
+        var headingPitchRoll = HeadingPitchRoll.fromQuaternion(quaternion, undefined, true);
+        var expected = new Matrix3(cPiOver2, -sPiOver2, 0.0, sPiOver2, cPiOver2, 0.0, 0.0, 0.0, 1.0);
+        var result = new Matrix3();
+        var returnedResult = Matrix3.fromHeadingPitchRoll(headingPitchRoll, result, true);
+        expect(result).toBe(returnedResult);
+        expect(returnedResult).toEqualEpsilon(expected, CesiumMath.EPSILON15);
+    });
+
+    it('fromHeadingPitchRoll computed correctly', function() {
+        // Expected generated via STK Components
+        var expected = new Matrix3(
+            0.754406506735489, 0.418940943945763, 0.505330889696038,
+            0.133022221559489, 0.656295369162553, -0.742685314912828,
+            -0.642787609686539, 0.627506871597133, 0.439385041770705);
+
+        var headingPitchRoll = new HeadingPitchRoll(CesiumMath.toRadians(10), CesiumMath.toRadians(40), CesiumMath.toRadians(55));
+        var result = new Matrix3();
+        var returnedResult = Matrix3.fromHeadingPitchRoll(headingPitchRoll, result, true);
+        expect(result).toBe(returnedResult);
+        expect(returnedResult).toEqualEpsilon(expected, CesiumMath.EPSILON15);
+    });
+
 
     it('fromScale works without a result parameter', function() {
         var expected = new Matrix3(